+++ src/components/common/report/Heading2.tsx
@@ -1,0 +1,5 @@
+const Heading2 = ({ children }: { children: React.ReactNode }) => {
+  return <h2 className="text-xsmall16 font-semibold">{children}</h2>;
+};
+
+export default Heading2;
