+++ src/components/common/report/Heading3.tsx
@@ -1,0 +1,7 @@
+const Heading3 = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <h3 className="mb-3 px-2.5 text-xsmall14 font-semibold">{children}</h3>
+  );
+};
+
+export default Heading3;
