+++ src/pages/common/report/ReportApplyPage.tsx
@@ -1,43 +1,15 @@
-import {
-  FormControl,
-  FormControlLabel,
-  InputLabel,
-  MenuItem,
-  Radio,
-  RadioGroup,
-  Select,
-} from '@mui/material';
-import { DatePicker } from '@mui/x-date-pickers';
+import { FormControl, RadioGroup } from '@mui/material';
+
+import ControlLabel from '../../../components/common/report/ControlLabel';
+import DateTimePicker from '../../../components/common/report/DateTimePicker';
+import FilledInput from '../../../components/common/report/FilledInput';
+import Heading2 from '../../../components/common/report/Heading2';
+import Heading3 from '../../../components/common/report/Heading3';
+import OutlinedButton from '../../../components/common/report/OutlinedButton';
+import Tooltip from '../../../components/common/report/Tooltip';
 
 const programName = '포트폴리오 조지기';
 const type = '포트폴리오';
-const timeOptions = [10, 11, 12, 13, 14];
-const radioSx = {
-  color: '#E7E7E7',
-  '&.Mui-checked': {
-    color: '#5177FF',
-  },
-};
-const labelSx = {
-  '.MuiFormControlLabel-label': {
-    fontSize: '0.875rem',
-    fontWeight: 500,
-    color: '#4C4F56',
-  },
-};
-const dateTimePickerSx = {
-  width: '50%',
-  label: {
-    fontSize: '0.875rem',
-    color: '#4C4F56',
-  },
-  '.MuiInputBase-formControl': {
-    borderRadius: '0.75rem',
-    fontSize: '0.875rem',
-    color: '#5177FF',
-    fontWeight: 500,
-  },
-};
 
 const ReportApplyPage = () => {
   return (
@@ -114,7 +86,7 @@ const DocumentSection = () => {
               value="file"
               subText="(pdf, doc, docx 형식 지원)"
             />
-            <OutlineButton caption="파일 업로드" />
+            <OutlinedButton caption="파일 업로드" />
           </div>
           {/* URL */}
           <div>
@@ -150,7 +122,7 @@ const PremiumSection = () => {
             <span className="-mt-1 mb-2 block text-xxsmall12 text-neutral-45">
               *업무, 지원자격, 우대사항이 보이게 채용공고를 캡처해주세요.
             </span>
-            <OutlineButton caption="파일 업로드" />
+            <OutlinedButton caption="파일 업로드" />
           </div>
           <div>
             <ControlLabel label="URL" value="url" />
@@ -214,94 +186,6 @@ const AdditionalInfoSection = () => {
   );
 };
 
-const Heading2 = ({ children }: { children: React.ReactNode }) => {
-  return <h2 className="text-xsmall16 font-semibold">{children}</h2>;
-};
-
-const Heading3 = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <h3 className="mb-3 px-2.5 text-xsmall14 font-semibold">{children}</h3>
-  );
-};
-
 const RequiredStar = () => {
   return <span className="text-[#7B61FF]">*</span>;
-};
-
-const Tooltip = ({ alt }: { alt: string }) => {
-  return (
-    <img
-      className="h-auto w-5 cursor-pointer"
-      src="/icons/message-question-circle.svg"
-      alt={alt}
-    />
-  );
-};
-
-const ControlLabel = ({
-  label,
-  value,
-  subText,
-}: {
-  label: string;
-  value: string;
-  subText?: string;
-}) => {
-  return (
-    <div className="flex items-center">
-      <FormControlLabel
-        sx={labelSx}
-        value={value}
-        control={<Radio size="small" sx={radioSx} />}
-        label={label}
-      />
-      {subText && (
-        <span className="-ml-2 inline-block text-xsmall14 font-medium text-neutral-50">
-          {subText}
-        </span>
-      )}
-    </div>
-  );
-};
-
-const OutlineButton = ({ caption }: { caption: string }) => {
-  return (
-    <button className="rounded-md border border-neutral-60 px-3 py-1.5 text-xsmall14 text-neutral-40">
-      {caption}
-    </button>
-  );
-};
-
-const FilledInput = ({ placeholder }: { placeholder?: string }) => {
-  return (
-    <div className="flex items-center">
-      <input
-        className="w-full rounded-md bg-neutral-95 p-3 text-xsmall14"
-        type="text"
-        placeholder={placeholder}
-      />
-    </div>
-  );
-};
-
-const DateTimePicker = () => {
-  return (
-    <div className="flex items-center gap-4">
-      <DatePicker
-        sx={dateTimePickerSx}
-        format="YY년 M월 D일(dd)"
-        label="날짜 선택"
-      />
-      <FormControl sx={dateTimePickerSx}>
-        <InputLabel id="time-select-label">시간 선택</InputLabel>
-        <Select labelId="time-select-label" label="시간 선택">
-          {timeOptions.map((option) => (
-            <MenuItem key={option} value={option}>
-              {`${option < 12 ? '오전' : '오후'} ${option < 13 ? option : option - 12}:00`}
-            </MenuItem>
-          ))}
-        </Select>
-      </FormControl>
-    </div>
-  );
 };
